working on display of network
