+++ src/core/frameworks/ethers/Web3Provider.ts
@@ -1,6 +1,6 @@
 import { JsonRpcProvider, JsonRpcSigner } from '@ethersproject/providers';
 
-import { getProviderType, getNetworkRpc, getNetworkId, Context } from '@utils';
+import { getProviderType, getNetworkRpc, getNetworkId } from '@utils';
 import { Network, ProviderType, Web3Provider, newSdkNetwork } from '@types';
 import { getConfig } from '@config';
 import { DebtDAO } from '@utils';
@@ -12,10 +12,9 @@ export class EthersWeb3ProviderImpl implements Web3Provider {
   private instances_1: Map<Network, DebtDAO<newSdkNetwork>> = new Map<Network, DebtDAO<newSdkNetwork>>();
 
   constructor() {
-    const { SUPPORTED_NETWORKS, CUSTOM_PROVIDER_HTTPS, USE_MAINNET_FORK } = getConfig();
-    SUPPORTED_NETWORKS.forEach((network) => {
+    const { DEBT_DAO_NETWORKS, CUSTOM_PROVIDER_HTTPS, USE_MAINNET_FORK } = getConfig();
+    DEBT_DAO_NETWORKS.forEach((network) => {
       const rpcUrl = getNetworkRpc(network);
-      const context = Context;
       const provider = getJsonRpcProvider(rpcUrl);
       const providerType = getProviderType(network);
       const networkId = getNetworkId(network) as newSdkNetwork;
@@ -27,6 +26,7 @@ export class EthersWeb3ProviderImpl implements Web3Provider {
       this.registerNetwork(network, newSdkNetwork);
     });
     if (USE_MAINNET_FORK) this.register('custom', getJsonRpcProvider(CUSTOM_PROVIDER_HTTPS));
+    console.log(DEBT_DAO_NETWORKS, 'TEST WEB3 PROVIDER NETWORKS');
   }
 
   public hasInstanceOf(type: ProviderType): boolean {
