+++ src/core/types/Config.ts
@@ -32,6 +32,7 @@ export interface Constants {
   DEBT_DAO_API: string;
   DEBT_DAO_ALERTS_API: string;
   SUPPORTED_NETWORKS: Network[];
+  DEBT_DAO_NETWORKS: Network[];
   NETWORK_SETTINGS: NetworkSettings;
   WEB3_PROVIDER_HTTPS: string;
   WEB3_PROVIDER_WSS: string;
